Make file import utils header-only to fix test target linking
